pyproject.toml
[project]
name = "bbb-image-forge"
version = "0.1.0"
description = "Beginner-friendly BeagleBone Black image builder"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "jinja2==3.1.6",
  "pydantic==2.11.7",
  "PyYAML==6.0.2",
  "requests==2.32.4",
  "beautifulsoup4==4.13.4",
  "typer==0.16.1",
  "python-multipart==0.0.20",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
